Three lists
Three lists
Given three lists of integers in lines separated by spaces.

The first line is the first list, the second is the second, the third is the third.

From the first list you need to select positive elements, from the second - even, from the third - multiples of 7. All selected elements must be combined into a single list and displayed in non-descending order.